runtime: Add online-kbs to kernel params
In online-kbs attestation the guest is given the location of the
keybroker server to connect after launch. This patch appends the
IP:Port of the online-kbs to the kernel params of the guest.

Patch also simplifies the kbs config into "mode" = offline/online,
and updates SEV config variable names and default values

Fixes: #5661 #5715

Signed-off-by: Jim Cadden <jcadden@ibm.com>