runtime-rs:add identification in version for runtime-rs
Since we now have 2 runtime, go runtime and rust runtime, and sometimes
it's hard for identifying which version of runtime being used.

In order to support debug runtime, we add identification for runtime-rs
version command (containerd-shim-kata-v2 --version) and an extra (rust)
will be displayed in --version when using rust runtime.

fixes:#5806

Signed-off-by: Chao Wu <chaowu@linux.alibaba.com>