runtime-rs: Introduce directly attachable network
As previously discussed at
https://github.com/kata-containers/kata-containers/issues/1922, the Kata
containers as VM-based containers are allowed to run in the host netns. That is,
the network is able to isolate in the L2. The network performance will benefit
from this architecture, which eliminates as many hops as possible. We called it
a Directly Attachable Network (DAN for short).

The network devices are placed at the host netns by the CNI plugins. The configs
are saved at `{dan_conf}/{sandbox_id}.json` in the format of JSON, including
device name, type, and network info. At a very beginning stage, the DAN only
supports host tap devices. More devices, like the DPDK, will be supported in
later versions. By the way, a CNI plugin named "dantap" could set up a bridge
and a tap device at the host netns. Please refer to
[dantap](https://github.com/justxuewei/cni-plugins/tree/feat/dan/plugins/main/dantap).

Fixes: #1922

Signed-off-by: Xuewei Niu <niuxuewei.nxw@antgroup.com>