agent: Use cgroup.kill where supported
Fixes: #6077

In 5.14+ you can write to a cgroup.kill file that will send a SIGKILL
to every process in the group for you. Support for this was recently
added to cgroups-rs and brought in here so we can make use of it. This
can avoid the Freeze -> Send signal -> Thaw dance that we use here, as
well as manually looping through the pids of a cgroup and sending SIGKILL
manually.

Signed-off-by: Danny Canter <danny@dcantah.dev>