runtime: unix socket file leak
When the shim process starts, a unix socket is created as the
service listening address, and it is created in the
/run/containerd/s/ directory. But it's not being cleaned up as
 expected when the shim process exits.

I found the corresponding cleaning logic in the shim framework code,
as follows:

// NOTE: If the shim server is down(like oom killer), the address
// socket might be leaking.
if address, err := ReadAddress("address"); err == nil {
	_ = RemoveSocket(address)
}

But because the shim process will exit through os.Exit(0)
in the ShutDown interface, the above code will not be executed.
We need to execute it before os.Exit(0).

Fixes: #6171

Signed-off-by: ls <335814617@qq.com>