kata-types/config: Add more default config paths
This is not ideal, but this is what it is right now, as the
configurations end up being loaded from that list in case
`KATA_CONF_FILE` is not set.

As part of our CI, as part of kata-deploy, and as part of the payload we
provide to Confidential Containers, we do not set `KATA_CONF_FILE`, thus
expanding the list seems like the most reasonable approach to take, at
least till github.com/kata-containers/kata-containers#3961 gets solved.

While expanding the list, let's also make sure to add the
`-dragonball.toml` as first options as nothing guarantees that the
default configuration will be pointing to that one.

Fixes: #6291

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>