utils: replaced jq with simple python3 in kata-manager
Most distributions do not have jq installed by default. Rather then
have a dependency that is only needed for the installation script and is
otherwise unlikely to be used, we can replace our use of jq for a small
embedded python program, something that is default on most systems.

Fixes #6270

Signed-off-by: Gabe Venberg <gabevenberg@gmail.com>