osbuilder: Support attestation-agent[cc_kbc] TDX dependencies
attestation-agent depends on tdx-attest-rs when cc_kbc is enabled, which
depends on libtdx-attest.so. Include the dev package in build container,
and the runtime package in the built rootfs.

The build of tdx-attest-sys (which is a dep of tdx-attest-rs) uses
bindgen, which requires libclang so install that in the build container
as well. There is also currently a bug in the build of tdx-attest-sys
crate when SGX_SDK is not provided in the env, so define that var as a
workaround. The bug is being fixed here:
https://github.com/intel/SGXDataCenterAttestationPrimitives/pull/288.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>