tools: allow multiple KBCs to be specified in AA_KBC
Right now each CoCo platform builds a separate initrd, with
attestation-agent configured to support a single KBC. It would be great
to eventually end up with a single generic initrd, where
attestation-agent supports multiple KBCs. As a first step in that
direction, enhance all checks for AA_KBC to support matching when
multiple KBCs are specified.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>