rust: Vendor all the dependencies
We've been hitting way too many errors trying to build the project and
failing when downloading the dependencies.  One way to solve this is
vendoring all our depdencies, regardless of this being common or not
with rust, it seems to be the most immediate solution for the baremetal
CIs we have.

Fixes: #6464

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>