gha: Split the Kata Containers CI workflow
We've been facing some issues in order to properly run, in a safe way,
the workflows for the Kata Containers CI.

This happens as the `pull_request` target doesn't allow passing any
secrets to workflows; and the `pull_request_target` workflow, which
allows passing the secrets, doesn't run on the PR itself but rather on
the target PR branch. :-/

https://securitylab.github.com/research/github-actions-preventing-pwn-requests/
proposes a reasonable solution, which I'm trying to mimic here.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>