dbs-boot: cache_info: define CacheType struct
The CacheType struct is used for distinguishing different types of
cache. There are three types at all: Data, Instruction and Unified.

Type Data and Instruction is used for level 1 cache, and type Unified is
used for non level 1 cache.

This commit implements the essential methods for CacheType struct, which
is used for creating fdt cache nodes.

Fixes: #6969

Signed-off-by: xuejun-xj <jiyunxue@linux.alibaba.com>