dbs-boot: cache_info: define CacheEngine struct
The CacheEngine struct manages the control flows of reading cache.

It set a limit to maximum threads used. According to this limitation,
the engine will calculate the suitable number of threads automatically
and use it to create new thread to read cache information concurrently.
Finally, receive the results and construct the HashMap to return all the
cache information to the caller.

It is created with a cpu mapping vector, which defines the mapping
relationship from vcpu(index) to pcpu(value).

Fixes: #6969

Signed-off-by: xuejun-xj <jiyunxue@linux.alibaba.com>