agent: add container launch control parameters from kernel commandline
`image_policy_file` is the URI of the image security policy file which
is used to control the allowlist and denylist of the containers to be
run inside the guest.

`image_registry_auth_file` is the URI of the auth file which is needed
to access an private registry.

`simple_signing_sigstore_config` is a configuration file that shows
where the signature files are when simple signing is enabled.

All the config items support both `kbs://..` scheme, `file://..` scheme,
or a direct absolute path `/...`, which means either the file is to be
fetched from a remote KBS that `aa_kbc_params` points to, or from the
local filesystem of the guest.

Fixes: #6640

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>