gha: Create ci-on-pr.yaml master workflow to run CI
Currently the only way to skip CI is by applying the force-skip-ci label manually.
By creating the ci-on-pr.yaml workflow to run on pull_requests and converting the
existing workflows to be callable we can automate the process of skipping the CI
for PRs where only static files ares modified (files under docs/*, *.png, etc).

The PR the workflows no longer have to check for each step whether the label is
present as each workflow will only be called when needed.

Credit to Hackathon Team3

Fixes: #3963
Signed-off-by: Tamas K Lengyel <tamas.lengyel@intel.com>